Add iTs for faiss
