+++ src/main/java/org/opensearch/knn/common/KNNConstants.java
@@ -44,6 +44,7 @@ public class KNNConstants {
 
     public static final String METHOD_HNSW = "hnsw";
 
+    public static final String METHOD_PARAMETER_EF_SEARCH = "ef_search";
     public static final String METHOD_PARAMETER_EF_CONSTRUCTION = "ef_construction";
     public static final String METHOD_PARAMETER_M = "m";
     public static final String METHOD_PARAMETER_SPACE_TYPE = "space_type"; // used for mapping parameter
